Improve js
