+++ deployment/frontend/public/sitemap-0.xml
@@ -1,18 +1,18 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>http://localhost:3000</loc><lastmod>2025-03-20T11:22:47.304Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>http://localhost:3000/applications</loc><lastmod>2025-03-20T11:22:47.304Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>http://localhost:3000/applications/404</loc><lastmod>2025-03-20T11:22:47.304Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>http://localhost:3000/auth/password-reset</loc><lastmod>2025-03-20T11:22:47.304Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>http://localhost:3000/datasets/404</loc><lastmod>2025-03-20T11:22:47.304Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>http://localhost:3000/search</loc><lastmod>2025-03-20T11:22:47.304Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>http://localhost:3000/teams</loc><lastmod>2025-03-20T11:22:47.304Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>http://localhost:3000/teams/404</loc><lastmod>2025-03-20T11:22:47.304Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>http://localhost:3000/topics</loc><lastmod>2025-03-20T11:22:47.304Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>http://localhost:3000/topics/404</loc><lastmod>2025-03-20T11:22:47.304Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>http://localhost:3000/user-guide</loc><lastmod>2025-03-20T11:22:47.304Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>http://localhost:3000/user-guide/basic-definition</loc><lastmod>2025-03-20T11:22:47.304Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>http://localhost:3000/user-guide/dataset-view-pages</loc><lastmod>2025-03-20T11:22:47.304Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>http://localhost:3000/user-guide/search-page</loc><lastmod>2025-03-20T11:22:47.304Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>http://localhost:3000/user-guide/teams-topics</loc><lastmod>2025-03-20T11:22:47.304Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:3000</loc><lastmod>2025-03-23T20:54:07.812Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:3000/applications</loc><lastmod>2025-03-23T20:54:07.812Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:3000/applications/404</loc><lastmod>2025-03-23T20:54:07.812Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:3000/auth/password-reset</loc><lastmod>2025-03-23T20:54:07.812Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:3000/datasets/404</loc><lastmod>2025-03-23T20:54:07.812Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:3000/search</loc><lastmod>2025-03-23T20:54:07.812Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:3000/teams</loc><lastmod>2025-03-23T20:54:07.812Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:3000/teams/404</loc><lastmod>2025-03-23T20:54:07.812Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:3000/topics</loc><lastmod>2025-03-23T20:54:07.812Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:3000/topics/404</loc><lastmod>2025-03-23T20:54:07.812Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:3000/user-guide</loc><lastmod>2025-03-23T20:54:07.812Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:3000/user-guide/basic-definition</loc><lastmod>2025-03-23T20:54:07.812Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:3000/user-guide/dataset-view-pages</loc><lastmod>2025-03-23T20:54:07.812Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:3000/user-guide/search-page</loc><lastmod>2025-03-23T20:54:07.812Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>http://localhost:3000/user-guide/teams-topics</loc><lastmod>2025-03-23T20:54:07.812Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
 </urlset>
